Installed dotenv, and added corresponding template.env, which will be useful for many future enhancements. package.json auto-formatted by npm. Moved error router out of app.js and user router out of routes/index.js and each into their own module. Added unifying index.js for models, as is the case for routes.
