+++ src/app.js
@@ -1,24 +1,13 @@
-var createError = require('http-errors');
+// Run dotenv as early as possible
+require('dotenv').config()
 var express = require('express');
 
-var indexRouter = require('./routes/index');
-
+// API Middleware Setup
 var app = express();
 app.use(express.json());
 
-app.use('/api', indexRouter);
-
-// catch 404 and forward to error handler
-app.use(function(req, res, next) {
-  next(createError(404));
-});
-
-// error handler
-app.use(function(err, req, res, next) {
-  // render the error page
-  res.status(err.status || 500);
-  res.send(err);
-});
+// Service Routes
+app.use('/api', require('./routes'));
 
 app.listen(3000, function() {
   console.log("Listening on port 3000");
