+++ src/models/index.js
@@ -1,0 +1,5 @@
+var User = require('./user');
+
+module.exports = {
+  User
+};
