+++ src/routes/index.js
@@ -1,23 +1,12 @@
-var User = require('../models/User');
-
 var express = require('express');
+var errorRouter = require('./error');
+var userRouter = require('./user');
+
+// Instantiate and apply sub-routers
 var router = express.Router();
+router.use('/user', userRouter);
 
-/* GET */
-router.get('/', function(req, res, next) {
-  res.send('Hello world!');
-});
-
-/* GET Users */
-router.get('/user', function(req, res, next) {
-  var users = [];
-  
-  users.push(new User({name: "John W"}));
-  users.push(new User({name: "Steph Z"}));
-  users.push(new User({name: "Cat"}));
-  users.push(new User({name: "Lily"}));
-  
-  res.send(users);
-});
+// Error handler must come last
+router.use(errorRouter);
 
 module.exports = router;
